# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ContentDetail" : {
        "properties" : {
          "authorDisplayName" : {
            "type" : "string"
          },
          "authorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "categories" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "commentCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "excerpt" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "markdownSource" : {
            "type" : "string"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "renderedHtml" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DRAFT", "APPROVED" ],
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "authorDisplayName", "categories", "commentCount", "createdAt", "id", "renderedHtml", "slug", "status", "tags", "title", "updatedAt" ],
        "type" : "object"
      },
      "ContentListResponse" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "posts" : {
            "items" : {
              "$ref" : "#/components/schemas/ContentSummary"
            },
            "type" : "array"
          },
          "totalCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "page", "posts", "totalCount", "totalPages" ],
        "type" : "object"
      },
      "ContentOperationConfirmation" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "message" ],
        "type" : "object"
      },
      "ContentSummary" : {
        "properties" : {
          "authorDisplayName" : {
            "type" : "string"
          },
          "categories" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "commentCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "excerpt" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "authorDisplayName", "categories", "commentCount", "id", "slug", "tags", "title" ],
        "type" : "object"
      },
...
      "CreateContentHttpRequest" : {
        "properties" : {
          "categoryIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array"
          },
          "markdownSource" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "markdownSource", "title" ],
        "type" : "object"
      },
      "CreateContentResponse" : {
        "properties" : {
          "authorDisplayName" : {
            "type" : "string"
          },
          "authorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "categories" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "excerpt" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DRAFT", "APPROVED" ],
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "authorDisplayName", "authorId", "categories", "createdAt", "id", "slug", "status", "tags", "title" ],
        "type" : "object"
      },
...
      "EditContentHttpRequest" : {
        "properties" : {
          "categoryIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array"
          },
          "markdownSource" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Content hub and knowledge management for the JVM ecosystem",
    "title" : "bytecode.news API",
    "version" : "0.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/posts/{id}" : {
      "delete" : {
        "operationId" : "deletePost",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "hard",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentOperationConfirmation"
                }
              }
            },
            "description" : "Post deleted"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Delete a post (soft by default, hard with ?hard=true)",
        "tags" : [ "Admin - Post Management" ]
      },
      "put" : {
        "operationId" : "editPost_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditContentHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post updated"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Admin edit a post",
        "tags" : [ "Admin - Post Management" ]
      }
    },
...
    "/posts" : {
      "get" : {
        "operationId" : "listPosts",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentListResponse"
                }
              }
            },
            "description" : "Paginated list of published posts"
          }
        },
        "summary" : "List published posts",
        "tags" : [ "Posts" ]
      },
      "post" : {
        "operationId" : "createPost",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateContentHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateContentResponse"
                }
              }
            },
            "description" : "Post created"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Create a new blog post draft",
        "tags" : [ "Posts" ]
      }
    },
    "/posts/search" : {
      "get" : {
        "operationId" : "searchPosts",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentListResponse"
                }
              }
            },
            "description" : "Search results"
          }
        },
        "summary" : "Search published posts",
        "tags" : [ "Posts" ]
      }
    },
    "/posts/{id}" : {
      "get" : {
        "operationId" : "getPostById",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post detail"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "summary" : "Get a post by ID",
        "tags" : [ "Posts" ]
      },
      "put" : {
        "operationId" : "editPost",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditContentHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post updated"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authorized to edit this post"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Edit a post",
        "tags" : [ "Posts" ]
      }
    },
    "/posts/{year}/{month}/{slug}" : {
      "get" : {
        "operationId" : "getPost",
        "parameters" : [ {
          "in" : "path",
          "name" : "year",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "maximum" : 3000,
            "minimum" : 2007,
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "month",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "maximum" : 12,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentDetail"
                }
              }
            },
            "description" : "Post detail"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Post not found"
          }
        },
        "summary" : "Get a published post by slug",
        "tags" : [ "Posts" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Local development server",
    "url" : "http://localhost:8080"
  } ]
}